Refactor layout running from POST:/api/document to document.applyLayout()
As before, the layout is run on a headless Cytoscape instance.  The final positions are applied to the document elements after the layout has finished running (async).  Thus, document.applyLayout() is mostly useful for serverside usecases.  For example, it could be useful as a clean-up step for a document that was created outside of UI or in a non-graph UI.