Avoid use of vw and vh units, especially for the 100vh case
iOS does not return the proper value for 100vh.  This change should make the page fit better on iOS devices, and it should prevent the case where the page does slight scrolling and gets stuck.  The scrolling bug can be reproduced without this change by swiping on the toolbar.  Sometimes popovers can reproduce the bug on show.