Exclude incomplete interactions from the submit button task list.
Those interactions are of type "other" and are as valid as automatically grounded entities.  They may not be as specific as they could be, but they are still valid.  For example, a user might not click any type, because "other" is the best fit and it's already selected by default.