GET /api/document : remove API key requirement & order by date
It is necessary for the home page exploration list to get the list of recent documents without the API key, because the home page is public.  In order to simplify things, the API key restriction has been removed:  The API key is not necessary in order to list the documents.  All documents are considered public information.

Further, the list of documents in the explore view should be ordered by most recent to least recent.  This ordering is applied by default by this route now.