Homepage redesign re. exploration & CTA
The homepage has been redesigned to allow for simple exploration of the published documents.  Above the fold, all documents in the published state are listed in the homepage's new horizontal scroller UI.  The scroller allows for navigation of the list of available documents using a familiar interface.

In addition, the overall design of the homepage has been simplified.  It now contains only two full-page views.  Below the fold, the project's aim and the prospective contributor's task are concisely defined.  The call to action is at the end:  With two simple pieces of information, the prospect's email and his article's title, the prospect can become a user.

Other design details:

- The banner images have been improved, applying darkening and gaussian blur to areas.
- There are new Contact, Github, and Twitter links.
- The page title is set depending on the view.  The prospect in the organic growth cycle would have two open tabs when he becomes a user.  This disambiguates them.
   - Home: `Biofactoid`
   - Editor or explore: `Article title : Biofactoid`
   - Admin: `Admin : Biofactoid`
- A sliding gradient animation is shown on a scroller document that is loading.

Other technical details:

- The document list route is amended to show only public information without the API key.  With the API key, secret information is returned.
- Requests for document snapshot thumbnails are cached in memory in an LRU cache.  At most `DOCUMENT_IMAGE_CACHE_SIZE` images are cached.  These images are cached as base 64 buffers.
- The scroller lists all documents when the app is run in debug mode.